+++ build_kernel_kenzo_miuiN.sh
@@ -15,8 +15,18 @@ then
 fi
 
 rm $KERNELDIR/arch/arm/boot/dts/*.dtb
-rm $KERNELDIR/drivers/staging/prima/wlan.ko
-rm $KERNELDIR/include/generated/compile.h
+if [ -f $KERNELDIR/drivers/staging/prima/wlan.ko ];
+	then
+	rm $KERNELDIR/drivers/staging/prima/wlan.ko
+fi
+if [ -f $KERNELDIR/drivers/staging/prima_n/wlan.ko ];
+	then
+	rm $KERNELDIR/drivers/staging/prima_n/wlan.ko
+fi
+if [ -f $KERNELDIR/include/generated/compile.h ];
+	then
+	rm $KERNELDIR/include/generated/compile.h
+fi
 if [ "`grep "AGNI_OREO_SYSMOUNT_MARKER" $KERNELDIR/arch/arm/boot/dts/qcom/kenzo/msm8956-kenzo.dtsi`" ];
 	then
 	if [ -f $KERNELDIR/arch/arm/boot/dts/qcom/kenzo/msm8956-kenzo.dtsi.bak ];
